moved swagger resource handler to swagger config
