+++ components/inspectit-ocelot-configurationserver/src/main/java/rocks/inspectit/ocelot/config/SwaggerConfiguration.java
@@ -3,6 +3,9 @@ package rocks.inspectit.ocelot.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.core.Authentication;
+import org.springframework.web.servlet.config.annotation.EnableWebMvc;
+import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import springfox.documentation.builders.PathSelectors;
 import springfox.documentation.builders.RequestHandlerSelectors;
 import springfox.documentation.spi.DocumentationType;
@@ -13,8 +16,9 @@ import springfox.documentation.swagger2.annotations.EnableSwagger2;
  * Configuration for enabling and providing swagger.
  */
 @Configuration
+@EnableWebMvc
 @EnableSwagger2
-public class SwaggerConfiguration {
+public class SwaggerConfiguration implements WebMvcConfigurer {
 
     @Bean
     public Docket swaggerApi() {
@@ -26,4 +30,13 @@ public class SwaggerConfiguration {
                 .build();
     }
 
+    @Override
+    public void addResourceHandlers(ResourceHandlerRegistry registry) {
+        registry
+                .addResourceHandler("swagger-ui.html")
+                .addResourceLocations("classpath:/META-INF/resources/");
+        registry
+                .addResourceHandler("/webjars/**")
+                .addResourceLocations("classpath:/META-INF/resources/webjars/");
+    }
 }
